refactor(core/dropdown): change import path
